Moved to resources
moved files to resources
