+++ resources/spotify.py
@@ -4,7 +4,9 @@ This file contains all the tools and functions needed for spotify integration.
 
 import spotipy
 from spotipy import oauth2
-from spotify import spotify_config
+from resources import spotify_config
+
+
 # from spotify_config import SPOTIFY_CLIENT_ID, SPOTIFY_CLIENT_SECRET, SPOTIFY_REDIRECT_URI, SCOPE, CACHE, \
 #    SONGBIRD_USER_ID
